Fix version bumping test packages to avoid package not publishing
Even in `test.pypi` if you try and publish a package with a version
already released, then PyPi will block you. Our current method for
determining the latest version uses the "releases" on GH, but for
test releases we don't create one, as I'm guessing we should only
create "releases" for production-ready code.

So using PyPi servers to grab the latest test version.

Also using `prerelease` instead of `prepatch` due to a bug in Poetry
which doesn't autoincrement the package number [0].

[0] https://github.com/python-poetry/poetry/issues/879